fix(auth): widen TOTP recovery-code length compare to u32
Pass 32 — crypto audit. The constant-time length compare in
`find_matching_recovery_code` cast both sides to u8, which mod-256
truncates: if FIXED_LEN ever grew past 255, a 0-byte input would
collide with a 256-byte stored code. Current recovery codes are
exactly 9 chars so this is defensive, not load-bearing, but the
function's docstring claims the length compare is correct for all
sizes — keep that promise by switching to u32.

The crypto audit was otherwise clean: subtle::ConstantTimeEq used
consistently for all secret/hash compares, AES-256-GCM with random
per-message nonces, OsRng for key/salt/nonce generation, JWT HS256
pinned (no algorithm confusion), `__Host-` / `__Secure-` cookie
prefixes, OAuth state via HMAC-binding + atomic GETDEL one-time
use, PKCE done correctly.

+++ crates/auth/src/totp.rs
@@ -81,8 +81,15 @@ pub fn find_recovery_code(codes: &[String], candidate: &str) -> Option<usize> {
 
     for (i, stored) in codes.iter().enumerate() {
         let stored_padded = pad_to_fixed(stored.as_bytes());
-        // Both length and content are compared in constant time
-        let len_match = (stored.len() as u8).ct_eq(&(candidate_len as u8));
+        // Both length and content are compared in constant time. Use u32
+        // for the length compare so it stays correct even if FIXED_LEN
+        // ever grows past 255 — `as u8` truncation would mod the length
+        // by 256 and turn e.g. a 0-byte input into a match for a 256-byte
+        // stored code. Current codes are exactly 9 chars so this is
+        // defensive, not load-bearing, but keeps the intent ("the length
+        // compare leaks no timing info AND is correct for all sizes")
+        // honest.
+        let len_match = (stored.len() as u32).ct_eq(&(candidate_len as u32));
         let content_match = stored_padded.ct_eq(&candidate_padded);
         if (len_match & content_match).into() {
             found_idx = Some(i);
